Seperating out the lambda protocol code
